feat(idea): resolve library dependency information
