+++ packages/cli/src/projects/ktjvm/src/main/kotlin/sample/main.kt
@@ -3,7 +3,7 @@ package sample
 import com.google.common.base.Joiner
 
 fun render_greeting(): String {
-  return Joiner.on(", ").join("Hello", "World")
+  return Joiner.on(" ").join("Hello", "World")
 }
 
 fun main() {
